[back] refactor : 인증 수단 전략 패턴의 key를 static 상수로 분리
